updating GitHub (xD)

+++ src/me/reckter/Generation/BackAndForth.java
@@ -1,9 +1,5 @@
 package me.reckter.Generation;
 
-import me.reckter.Util;
-
-import java.awt.color.ColorSpace;
-
 /**
  * Created with IntelliJ IDEA.
  * User: reckter
@@ -23,9 +19,9 @@ public class BackAndForth extends BasicGeneration {
 					x = (r * COLORS * COLORS + g * COLORS + b) % SIZE;
 					y = (int) ((float)(r * COLORS * COLORS + g * COLORS + b) / (float) SIZE);
 					//Util.c_log("(" + x + "|" + y + ")");
-					pixel[x][y][R] = (short) (b % 2 == 0 ? r : 255 - r);
-					pixel[x][y][G] = (short) (r % 2 == 0 ? g : 255 - g);
-					pixel[x][y][B] = (short) (g % 2 == 0 ? b : 255 - b);
+					pixel[x][y][R] = (byte) (b % 2 == 0 ? r : 255 - r);
+					pixel[x][y][G] = (byte) (r % 2 == 0 ? g : 255 - g);
+					pixel[x][y][B] = (byte) (g % 2 == 0 ? b : 255 - b);
 				}
 			}
 		}
